mpi_size_t -> mpi_int

+++ tasks/all/kholin_k_multidimensional_integrals_rectangle/func_tests/func_all.cpp
@@ -11,7 +11,6 @@
 #endif
 
 #include <cmath>
-#include <cstddef>
 #include <cstdint>
 #include <functional>
 #include <memory>
@@ -24,7 +23,7 @@ TEST(kholin_k_multidimensional_integrals_rectangle_all, test_validation) {
   int rank = 0;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   // Create data
-  size_t dim = 1;
+  int dim = 1;
   std::vector<double> values{0.0};
   auto f = [](const std::vector<double> &f_values) { return std::sin(f_values[0]); };
   std::vector<double> in_lower_limits{0};
@@ -56,7 +55,7 @@ TEST(kholin_k_multidimensional_integrals_rectangle_all, test_pre_processing) {
   int rank = 0;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   // Create data
-  size_t dim = 1;
+  int dim = 1;
   std::vector<double> values{0.0};
   auto f = [](const std::vector<double> &f_values) { return std::sin(f_values[0]); };
   std::vector<double> in_lower_limits{0};
@@ -89,7 +88,7 @@ TEST(kholin_k_multidimensional_integrals_rectangle_all, test_run) {
   int rank = 0;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   // Create data
-  size_t dim = 1;
+  int dim = 1;
   std::vector<double> values{0.0};
   auto f = [](const std::vector<double> &f_values) { return std::sin(f_values[0]); };
   std::vector<double> in_lower_limits{0};
@@ -123,7 +122,7 @@ TEST(kholin_k_multidimensional_integrals_rectangle_all, test_post_processing) {
   int rank = 0;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   // Create data
-  size_t dim = 1;
+  int dim = 1;
   std::vector<double> values{0.0};
   auto f = [](const std::vector<double> &f_values) { return std::sin(f_values[0]); };
   std::vector<double> in_lower_limits{0};
@@ -158,7 +157,7 @@ TEST(kholin_k_multidimensional_integrals_rectangle_all, single_integral_one_var)
   int rank = 0;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   // Create data
-  size_t dim = 1;
+  int dim = 1;
   std::vector<double> values{0.0};
   auto f = [](const std::vector<double> &f_values) { return f_values[0]; };
   std::vector<double> in_lower_limits{2};
@@ -198,7 +197,7 @@ TEST(kholin_k_multidimensional_integrals_rectangle_all, single_integral_two_var)
   int rank = 0;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   // Create data
-  size_t dim = 1;
+  int dim = 1;
   std::vector<double> values{0.0, 3.0};
   auto f = [](const std::vector<double> &f_values) { return f_values[0] + f_values[1]; };
   std::vector<double> in_lower_limits{0};
@@ -239,7 +238,7 @@ TEST(kholin_k_multidimensional_integrals_rectangle_all, double_integral_two_var)
   int rank = 0;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   // Create data
-  size_t dim = 2;
+  int dim = 2;
   std::vector<double> values{0.0, 0.0};
   auto f = [](const std::vector<double> &f_values) { return (2 * f_values[0]) + (2 * f_values[1]); };
   std::vector<double> in_lower_limits{0, 0};
@@ -280,7 +279,7 @@ TEST(kholin_k_multidimensional_integrals_rectangle_all, double_integral_one_var)
   int rank = 0;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   // Create data
-  size_t dim = 2;
+  int dim = 2;
   std::vector<double> values{-17.0, 0.0};
   auto f = [](const std::vector<double> &f_values) { return 289 + (f_values[1] * f_values[1]); };
   std::vector<double> in_lower_limits{-10, 3};
@@ -321,7 +320,7 @@ TEST(kholin_k_multidimensional_integrals_rectangle_all, triple_integral_three_va
   int rank = 0;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   // Create data
-  size_t dim = 3;
+  int dim = 3;
   std::vector<double> values{0.0, 0.0, 0.0};
   auto f = [](const std::vector<double> &f_values) {
     return (f_values[0] * f_values[0]) + (f_values[1] * f_values[1]) + (f_values[2] * f_values[2]);
@@ -364,7 +363,7 @@ TEST(kholin_k_multidimensional_integrals_rectangle_all, triple_integral_two_var)
   int rank = 0;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   // Create data
-  size_t dim = 3;
+  int dim = 3;
   std::vector<double> values{0.0, 5.0, 0.0};
   auto f = [](const std::vector<double> &f_values) { return (f_values[0] + f_values[1]); };
   std::vector<double> in_lower_limits{0, 0, 0};
@@ -405,7 +404,7 @@ TEST(kholin_k_multidimensional_integrals_rectangle_all, triple_integral_one_var)
   int rank = 0;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   // Create data
-  size_t dim = 3;
+  int dim = 3;
   std::vector<double> values{0.0, 5.0, -10.0};
   auto f = [](const std::vector<double> &f_values) { return f_values[0] + 5.0 + (-10.0); };
   std::vector<double> in_lower_limits{0, 0, 0};
